requirements.txt
gradio
httpx[http2]
trafilatura
python-dateutil
limits
filelock
sentence-transformers
transformers
faiss-cpu
torch
scikit-learn
accelerate
